feat: rate limit 테스트 코드 작성

+++ module-common/src/main/kotlin/com/example/redis/annotations/RateLimited.kt
@@ -4,5 +4,5 @@ package com.example.redis.annotations
 @Target(AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.RUNTIME)
 annotation class RateLimited(
-    val value: Double
+    val value: Long
 )
